run new test with Scala 2.13+ only
